requirements.txt
click
flask
PyJWT
flask_restful
flask_cors
